changed NUM type from int to long
this change required amazingly little work to effect:
* change %d to %ld in print.c
* change atoi to atol in parse.c
* change the Obj num type in objects.h

have I forgotten something? I can't think of anything.

+++ parse.c
@@ -139,7 +139,7 @@ Obj read_tokens(Token_list* tokens) {
 		// if (DEBUG) printf("token: %s\n", token.text);
 		char* text = token.text;
 		if (isdigit(text[0])) 
-			obj = NUMOBJ(atoi(text));
+			obj = NUMOBJ(atol(text));
 		else 
 			obj = NAMEOBJ(text);
 		return obj;
